[aiecc] Fold StackSizeAnalysis into tools/aiecc, drop lib/Analysis
lib/Analysis existed as a new top-level category alongside Dialect/
Conversion/Target purely for one file, with a single consumer (aiecc) and
no standalone unit test exercising it as a library in its own right --
it's only ever tested end-to-end through the aiecc driver. Its own name
also collides with the established convention in both LLVM and MLIR,
where lib/Analysis means IR-level analysis (AliasAnalysis, CallGraph,
Liveness); this parses relocatable object files via llvm::object, closer
in kind to llvm/lib/Object.

Move StackSizeAnalysis.{h,cpp} into tools/aiecc as plain driver-local
sources, matching Graph.h/Utils.h/ExecutionEngine.h/IRTransforms.h already
living there, and delete the now-empty lib/Analysis and its CMake target.
