+++ lib/CMakeLists.txt
@@ -5,7 +5,6 @@
 #
 
 
-add_subdirectory(Analysis)
 add_subdirectory(Targets)
 add_subdirectory(Dialect)
 add_subdirectory(CAPI)
